requirements.txt
streamlit
anthropic
python-docx
python-dotenv
PyPDF2